move deploy to lib
